Fix clang-tidy and compiler warnings across the codebase
This commit addresses various clang-tidy checks and compiler warnings:

- Sign conversion fixes: Added explicit casts for SQLLEN/size_t conversions
  in BasicStringBinder, SqlDynamicBinary, SqlDynamicString, SqlStatement,
  DataMapper, CxxModelPrinter, and libzip API calls
- Cognitive complexity: Added NOLINT comments for complex but necessary
  functions in BatchManager, SqlBackup, and DataGenerator
- Empty catch statements: Added NOLINT comments with explanatory notes
  for intentional empty catches in SqlSchema and SqlBackup
- Modernize-use-auto: Changed explicit types to auto where appropriate
- Modernize-use-ranges: Changed std::find_if/any_of to ranges versions
- Missing std::forward: Added NOLINT for intentionally not forwarding
- Uninitialized struct members: Added default initializers
- HeaderFilterRegex: Updated to exclude libzip-src from analysis

All tests pass with clang-tidy and asan/ubsan enabled.

+++ src/Lightweight/DataBinder/SqlDynamicBinary.hpp
@@ -202,20 +202,20 @@ struct SqlDataBinder<SqlDynamicBinary<N>>
                 // We have a truncation and the server does not know how much data is left.
                 result->resize(result->size() - 1);
             else if (*indicator <= static_cast<SQLLEN>(result->size()))
-                result->resize(*indicator);
+                result->resize(static_cast<size_t>(*indicator));
             else
             {
                 // We have a truncation and the server knows how much data is left.
                 // Extend the buffer and fetch the rest via SQLGetData.
 
                 auto const totalCharsRequired = *indicator;
-                result->resize(totalCharsRequired + 1);
+                result->resize(static_cast<size_t>(totalCharsRequired + 1));
                 auto const sqlResult =
                     SQLGetData(stmt, column, SQL_C_BINARY, (SQLPOINTER) result->data(), totalCharsRequired + 1, indicator);
                 (void) sqlResult;
                 assert(SQL_SUCCEEDED(sqlResult));
                 assert(*indicator == totalCharsRequired);
-                result->resize(totalCharsRequired);
+                result->resize(static_cast<size_t>(totalCharsRequired));
             }
         });
